Data_Science_Projects/Autonomous_Vehicle_Vision_EV_Analytics/requirements.txt
# Autonomous Vehicle Vision & EV Sensor Analytics - Requirements

# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Computer Vision & Deep Learning
opencv-python>=4.8.0
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8
Pillow>=10.0.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.10.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Web Framework & API
flask>=2.3.0
flask-cors>=4.0.0
flask-restful>=0.3.10
gunicorn>=21.2.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL
mysql-connector-python>=8.0.0

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# Data Processing & Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0

# Visualization
plotly>=5.17.0
dash>=2.14.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0

